kafka/00-setup/kafka_basic/requirements.txt
confluent-kafka==2.13.2
dnspython==2.7.0
pymongo==4.16.0
python-dotenv==1.2.1
